backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.0
python-dotenv==1.0.1
google-generativeai==0.3.2
Pillow==10.2.0
python-multipart==0.0.7
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2
pymongo==4.6.1
email-validator==2.1.1
